Move shared functions and constants to global language data
