+++ spacy/de/language_data.py
@@ -2,16 +2,10 @@
 from __future__ import unicode_literals
 
 from ..symbols import *
+from ..language_data import PRON_LEMMA
 from ..language_data import TOKENIZER_PREFIXES
 from ..language_data import TOKENIZER_SUFFIXES
 from ..language_data import TOKENIZER_INFIXES
-
-
-def strings_to_exc(orths):
-    return {orth: [{ORTH: orth}] for orth in orths}
-
-
-PRON_LEMMA = "-PRON-"
 
 
 TAG_MAP = {
