+++ spacy/en/__init__.py
@@ -12,9 +12,9 @@ from ..tokenizer import Tokenizer
 from ..attrs import LANG
 
 from ..language_data import update_exc
+from ..language_data import strings_to_exc
 from ..language_data import EMOTICONS
 from .language_data import ORTH_ONLY
-from .language_data import strings_to_exc
 from .language_data import get_time_exc
