+++ spacy/pt/__init__.py
@@ -8,9 +8,9 @@ from . import language_data
 from ..attrs import LANG
 
 from ..language_data import update_exc
+from ..language_data import strings_to_exc
 from ..language_data import EMOTICONS
 from .language_data import ORTH_ONLY
-from .language_data import strings_to_exc
 
 
 TOKENIZER_EXCEPTIONS = dict(language_data.TOKENIZER_EXCEPTIONS)
